backend/requirements.txt
fastapi
uvicorn[standard]
pyspellchecker==0.6.2
requests
python-dotenv
